src/main/webapp/app/entities/service-mfpai/route/service-mfpai-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IServiceMFPAI, ServiceMFPAI } from '../service-mfpai.model';
import { ServiceMFPAIService } from '../service/service-mfpai.service';

import { ServiceMFPAIRoutingResolveService } from './service-mfpai-routing-resolve.service';

describe('ServiceMFPAI routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: ServiceMFPAIRoutingResolveService;
  let service: ServiceMFPAIService;
  let resultServiceMFPAI: IServiceMFPAI | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(ServiceMFPAIRoutingResolveService);
    service = TestBed.inject(ServiceMFPAIService);
    resultServiceMFPAI = undefined;
  });

  describe('resolve', () => {
    it('should return IServiceMFPAI returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultServiceMFPAI = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultServiceMFPAI).toEqual({ id: 123 });
    });

    it('should return new IServiceMFPAI if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultServiceMFPAI = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultServiceMFPAI).toEqual(new ServiceMFPAI());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as ServiceMFPAI })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultServiceMFPAI = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultServiceMFPAI).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
